+++ NEWS.md
@@ -1,9 +1,10 @@
 # dials 0.0.1.9000
 
-* `regularization` was changed to `penalty` in a few models to be consistent with [this change](tidymodels/model-implementation-principles@08d3afd). 
 * Parameter objects now contain code to finalize their values and a number of helper functions for certain data-specific parameters. A `force` option can be used to avoid updating the values.  
 * Parameter objects are printed differently inside of tibbles. 
+* `regularization` was changed to `penalty` in a few models to be consistent with [this change](tidymodels/model-implementation-principles@08d3afd). 
 * `batch_size` and `threshold` were added.
+* Added a set of parameters for the `textrecipes` package [issue 16](https://github.com/tidymodels/dials/issues/16). 
 
 # dials 0.0.1